Reduced size of test.
